requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
python-multipart==0.0.6
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.26.0
python-barcode>=0.15.1
Pillow>=10.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
wheel>=0.42.0

pytest>=7.4.4
httpx>=0.26.0
coverage>=7.0.0
flake8>=7.0.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9

